+++ src/Servers/Kestrel/Transport.MsQuic/src/Internal/MsQuicConnection.cs
@@ -2,44 +2,56 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
-using System.Collections.Generic;
+using System.Runtime.InteropServices;
 using System.Threading.Channels;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Connections;
 using Microsoft.AspNetCore.Connections.Abstractions.Features;
 using Microsoft.AspNetCore.Connections.Features;
+using static Microsoft.AspNetCore.Server.Kestrel.Transport.MsQuic.Internal.MsQuicNativeMethods;
 
 namespace Microsoft.AspNetCore.Server.Kestrel.Transport.MsQuic.Internal
 {
-    internal class MsQuicConnection : TransportConnection, IQuicStreamListenerFeature, IQuicCreateStreamFeature
+    internal class MsQuicConnection : TransportConnection, IQuicStreamListenerFeature, IQuicCreateStreamFeature, IDisposable
     {
-        private QuicConnection _connection;
+        private bool _disposed;
+        private readonly MsQuicTransportContext _context;
+        private IntPtr _nativeObjPtr;
+        private ConnectionCallback _callback;
+        private static GCHandle _handle;
+        private readonly IntPtr _unmanagedFnPtrForNativeCallback;
+        private TaskCompletionSource<object> _tcsConnection;
+
         private readonly Channel<MsQuicStream> _acceptQueue = Channel.CreateUnbounded<MsQuicStream>(new UnboundedChannelOptions
         {
             SingleReader = true,
             SingleWriter = true
         });
 
-        public MsQuicConnection(QuicConnection connection)
+        public MsQuicConnection(MsQuicApi registration, MsQuicTransportContext context, IntPtr nativeObjPtr)
         {
-            _connection = connection;
-            connection.SetCallbackHandler(HandleEvent);
-            connection.SetIdleTimeout(TimeSpan.FromMinutes(10));
+            Registration = registration;
+            _context = context;
+            _nativeObjPtr = nativeObjPtr;
+
+            ConnectionCallbackDelegate nativeCallback = NativeCallbackHandler;
+            _unmanagedFnPtrForNativeCallback = Marshal.GetFunctionPointerForDelegate(nativeCallback);
+            SetCallbackHandler(HandleEvent);
+            SetIdleTimeout(TimeSpan.FromMinutes(10));
 
             Features.Set<IQuicStreamListenerFeature>(this);
+            Features.Set<IQuicCreateStreamFeature>(this);
         }
 
         public uint HandleEvent(
-            QuicConnection connection,
             ref MsQuicNativeMethods.ConnectionEvent evt)
         {
-            var status = HandleEventCore(connection, ref evt);
+            var status = HandleEventCore(ref evt);
 
             return status;
         }
 
         private uint HandleEventCore(
-          QuicConnection connection,
           ref MsQuicNativeMethods.ConnectionEvent connectionEvent)
         {
             var status = MsQuicConstants.Success;
@@ -48,7 +60,6 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Transport.MsQuic.Internal
                 case QUIC_CONNECTION_EVENT.CONNECTED:
                     {
                         status = HandleEventConnected(
-                            connection,
                             connectionEvent);
                     }
                     break;
@@ -56,7 +67,6 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Transport.MsQuic.Internal
                 case QUIC_CONNECTION_EVENT.SHUTDOWN_BEGIN:
                     {
                         status = HandleEventShutdownBegin(
-                            connection,
                             connectionEvent);
                     }
                     break;
@@ -64,7 +74,6 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Transport.MsQuic.Internal
                 case QUIC_CONNECTION_EVENT.SHUTDOWN_BEGIN_PEER:
                     {
                         status = HandleEventShutdownBeginPeer(
-                            connection,
                             connectionEvent);
                     }
                     break;
@@ -72,7 +81,6 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Transport.MsQuic.Internal
                 case QUIC_CONNECTION_EVENT.SHUTDOWN_COMPLETE:
                     {
                         status = HandleEventShutdownComplete(
-                            connection,
                             connectionEvent);
                     }
                     break;
@@ -80,7 +88,6 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Transport.MsQuic.Internal
                 case QUIC_CONNECTION_EVENT.LOCAL_ADDR_CHANGED:
                     {
                         status = HandleEventLocalAddrChanged(
-                            connection,
                             connectionEvent);
                     }
                     break;
@@ -89,7 +96,6 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Transport.MsQuic.Internal
                 case QUIC_CONNECTION_EVENT.PEER_ADDR_CHANGED:
                     {
                         status = HandleEventPeerAddrChanged(
-                            connection,
                             connectionEvent);
                     }
                     break;
@@ -97,7 +103,6 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Transport.MsQuic.Internal
                 case QUIC_CONNECTION_EVENT.NEW_STREAM:
                     {
                         status = HandleEventNewStream(
-                            connection,
                             connectionEvent);
                     }
                     break;
@@ -105,7 +110,6 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Transport.MsQuic.Internal
                 case QUIC_CONNECTION_EVENT.STREAMS_AVAILABLE:
                     {
                         status = HandleEventStreamsAvailable(
-                            connection,
                             connectionEvent);
                     }
                     break;
@@ -118,7 +122,6 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Transport.MsQuic.Internal
                 case QUIC_CONNECTION_EVENT.IDEAL_SEND_BUFFER:
                     {
                         status = HandleEventIdealSendBuffer(
-                            connection,
                             connectionEvent);
                     }
                     break;
@@ -130,69 +133,59 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Transport.MsQuic.Internal
         }
 
         protected virtual uint HandleEventConnected(
-            QuicConnection connection,
             MsQuicNativeMethods.ConnectionEvent connectionEvent)
         {
             return MsQuicConstants.Success;
         }
 
         protected virtual uint HandleEventShutdownBegin(
-            QuicConnection connection,
             MsQuicNativeMethods.ConnectionEvent connectionEvent)
         {
             return MsQuicConstants.Success;
         }
 
         protected virtual uint HandleEventShutdownBeginPeer(
-            QuicConnection connection,
             MsQuicNativeMethods.ConnectionEvent connectionEvent)
         {
             return MsQuicConstants.Success;
         }
 
         protected virtual uint HandleEventShutdownComplete(
-            QuicConnection connection,
             MsQuicNativeMethods.ConnectionEvent connectionEvent)
         {
             return MsQuicConstants.Success;
         }
 
         protected virtual uint HandleEventLocalAddrChanged(
-            QuicConnection connection,
             MsQuicNativeMethods.ConnectionEvent connectionEvent)
         {
             return MsQuicConstants.Success;
         }
 
         protected virtual uint HandleEventPeerAddrChanged(
-            QuicConnection connection,
             MsQuicNativeMethods.ConnectionEvent connectionEvent)
         {
             return MsQuicConstants.Success;
         }
 
         protected virtual uint HandleEventNewStream(
-            QuicConnection connection,
             MsQuicNativeMethods.ConnectionEvent connectionEvent)
         {
-            var stream = new QuicStream(connection.Registration, connectionEvent.Data.NewStream.Stream, shouldOwnNativeObj: false);
-            var streamWrapper = new MsQuicStream(this, connectionEvent.StreamFlags);
+            //var stream = new QuicStream(connection.Registration, connectionEvent.Data.NewStream.Stream, shouldOwnNativeObj: false);
+            var msQuicStream = new MsQuicStream(Registration, this, _context, connectionEvent.StreamFlags, connectionEvent.Data.NewStream.Stream);
 
-            streamWrapper.Start(stream);
-            _acceptQueue.Writer.TryWrite(streamWrapper);
+            _acceptQueue.Writer.TryWrite(msQuicStream);
 
             return MsQuicConstants.Success;
         }
 
         protected virtual uint HandleEventStreamsAvailable(
-            QuicConnection connection,
             MsQuicNativeMethods.ConnectionEvent connectionEvent)
         {
             return MsQuicConstants.Success;
         }
 
         protected virtual uint HandleEventIdealSendBuffer(
-            QuicConnection connection,
             MsQuicNativeMethods.ConnectionEvent connectionEvent)
         {
             return MsQuicConstants.Success;
@@ -223,11 +216,210 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Transport.MsQuic.Internal
 
         private async ValueTask<ConnectionContext> StartStreamAsync(QUIC_STREAM_OPEN_FLAG flags)
         {
-            var msquicStream = new MsQuicStream(this, flags);
-            var stream = _connection.StreamOpen(flags, msquicStream.HandleStreamEvent);
+            var stream = StreamOpen(flags);
             await stream.StartAsync(QUIC_STREAM_START_FLAG.NONE);
-            msquicStream.Start(stream);
-            return msquicStream;
+            return stream;
+        }
+
+
+        public MsQuicApi Registration { get; set; }
+
+        public delegate uint ConnectionCallback(
+            ref ConnectionEvent connectionEvent);
+
+
+        public unsafe void SetIdleTimeout(TimeSpan timeout)
+        {
+            var msTime = (ulong)timeout.TotalMilliseconds;
+            var buffer = new QuicBuffer()
+            {
+                Length = sizeof(ulong),
+                Buffer = (byte*)&msTime
+            };
+            SetParam(QUIC_PARAM_CONN.IDLE_TIMEOUT, buffer);
+        }
+
+        public void SetPeerBiDirectionalStreamCount(ushort count)
+        {
+            SetUshortParamter(QUIC_PARAM_CONN.PEER_BIDI_STREAM_COUNT, count);
+        }
+
+        public void SetPeerUnidirectionalStreamCount(ushort count)
+        {
+            SetUshortParamter(QUIC_PARAM_CONN.PEER_UNIDI_STREAM_COUNT, count);
+        }
+
+        public void SetLocalBidirectionalStreamCount(ushort count)
+        {
+            SetUshortParamter(QUIC_PARAM_CONN.LOCAL_BIDI_STREAM_COUNT, count);
+        }
+
+        public void SetLocalUnidirectionalStreamCount(ushort count)
+        {
+            SetUshortParamter(QUIC_PARAM_CONN.LOCAL_UNIDI_STREAM_COUNT, count);
+        }
+
+        private unsafe void SetUshortParamter(QUIC_PARAM_CONN param, ushort count)
+        {
+            var buffer = new QuicBuffer()
+            {
+                Length = sizeof(ushort),
+                Buffer = (byte*)&count
+            };
+            SetParam(param, buffer);
+        }
+
+        public unsafe void EnableBuffering()
+        {
+            var val = true;
+            var buffer = new QuicBuffer()
+            {
+                Length = sizeof(bool),
+                Buffer = (byte*)&val
+            };
+            SetParam(QUIC_PARAM_CONN.USE_SEND_BUFFER, buffer);
+        }
+
+        public unsafe void DisableBuffering()
+        {
+            var val = false;
+            var buffer = new QuicBuffer()
+            {
+                Length = sizeof(bool),
+                Buffer = (byte*)&val
+            };
+            SetParam(QUIC_PARAM_CONN.USE_SEND_BUFFER, buffer);
+        }
+
+        public Task StartAsync(
+            ushort family,
+            string serverName,
+            ushort serverPort)
+        {
+            _tcsConnection = new TaskCompletionSource<object>();
+
+            var status = Registration.ConnectionStartDelegate(
+                _nativeObjPtr,
+                family,
+                serverName,
+                serverPort);
+
+            MsQuicStatusException.ThrowIfFailed(status);
+
+            return _tcsConnection.Task;
+        }
+
+        public MsQuicStream StreamOpen(
+            QUIC_STREAM_OPEN_FLAG flags)
+        {
+            var streamPtr = IntPtr.Zero;
+            var status = Registration.StreamOpenDelegate(
+                _nativeObjPtr,
+                (uint)flags,
+                MsQuicStream.NativeCallbackHandler,
+                IntPtr.Zero,
+                out streamPtr);
+            MsQuicStatusException.ThrowIfFailed(status);
+
+            return new MsQuicStream(Registration, this, _context, flags, streamPtr);
+        }
+
+        public void SetCallbackHandler(
+            ConnectionCallback callback)
+        {
+            _handle = GCHandle.Alloc(this);
+            _callback = callback;
+            Registration.SetCallbackHandlerDelegate(
+                _nativeObjPtr,
+                _unmanagedFnPtrForNativeCallback,
+                GCHandle.ToIntPtr(_handle));
+        }
+
+        public void Shutdown(
+            QUIC_CONNECTION_SHUTDOWN_FLAG Flags,
+            ushort ErrorCode)
+        {
+            var status = Registration.ConnectionShutdownDelegate(
+                _nativeObjPtr,
+                (uint)Flags,
+                ErrorCode);
+            MsQuicStatusException.ThrowIfFailed(status);
+        }
+
+        internal static uint NativeCallbackHandler(
+            IntPtr connection,
+            IntPtr context,
+            ref ConnectionEvent connectionEventStruct)
+        {
+            var handle = GCHandle.FromIntPtr(context);
+            var quicConnection = (MsQuicConnection)handle.Target;
+            return quicConnection.ExecuteCallback(ref connectionEventStruct);
+        }
+
+        private uint ExecuteCallback(
+            ref ConnectionEvent connectionEvent)
+        {
+            var status = MsQuicConstants.InternalError;
+            if (connectionEvent.Type == QUIC_CONNECTION_EVENT.CONNECTED)
+            {
+                _tcsConnection?.TrySetResult(null);
+            }
+            else if (connectionEvent.Type == QUIC_CONNECTION_EVENT.SHUTDOWN_BEGIN)
+            {
+                _tcsConnection?.TrySetResult(new Exception("RIP"));
+            }
+
+            try
+            {
+                status = _callback(
+                    ref connectionEvent);
+            }
+            catch (Exception)
+            {
+                // TODO log
+            }
+            return status;
+        }
+
+        private void SetParam(
+            QUIC_PARAM_CONN param,
+            QuicBuffer buf)
+        {
+            MsQuicStatusException.ThrowIfFailed(Registration.UnsafeSetParam(
+                _nativeObjPtr,
+                (uint)QUIC_PARAM_LEVEL.CONNECTION,
+                (uint)param,
+                buf));
+        }
+
+        public void Dispose()
+        {
+            Dispose(true);
+            GC.SuppressFinalize(this);
+        }
+
+        ~MsQuicConnection()
+        {
+            Dispose(false);
+        }
+
+        private void Dispose(bool disposing)
+        {
+            if (_disposed)
+            {
+                return;
+            }
+
+            if (_nativeObjPtr != IntPtr.Zero)
+            {
+                Registration.ConnectionCloseDelegate?.Invoke(_nativeObjPtr);
+            }
+
+            _nativeObjPtr = IntPtr.Zero;
+            Registration = null;
+
+            _handle.Free();
+            _disposed = true;
         }
     }
 }
